pdf-agent/requirements.txt

pdfplumber
pdf2image
pillow
pytesseract
sentence-transformers
faiss-cpu
fastapi
uvicorn[standard]
streamlit
